+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+* Upgrade the `org.realityforge.arez.testng` artifact to version `0.24`.
 * Upgrade the `org.realityforge.arez` artifacts to version `0.192`.
 
 ### [v0.179](https://github.com/react4j/react4j/tree/v0.179) (2021-01-07) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.178...v0.179) · [API Differences](https://react4j.github.io/api-diff?key=react4j&old=0.178&new=0.179)